# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for tracking daily activities, streaks, and analytics.\n\n## Features\n- User authentication with JWT tokens\n- Activity tracking with customizable types\n- Streak tracking to maintain consistency\n- Weekly analytics with daily breakdowns\n- Privacy controls for user profiles\n- Tile configuration for dashboard customization\n\n## Rate Limits\n- Login/Register: 5 requests per minute per IP\n- Password Reset: 3 requests per minute per IP\n- General API: 60 requests per minute per user",
        "title": "Growth Tracker API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/aman1117/growth-tracker"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/create-activity": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Log hours for a specific activity on a date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activities"
                ],
                "summary": "Create or update an activity",
                "parameters": [
                    {
                        "description": "Activity details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.CreateActivityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity updated successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/get-activities": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve activities for a user within a date range",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activities"
                ],
                "summary": "Get activities for a user",
                "parameters": [
                    {
                        "description": "Date range and username",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.GetActivitiesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of activities",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.DataResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ActivityDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validation error or user not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Private account",
                        "schema": {
                            "$ref": "#/definitions/github_com_aman1117_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_aman1117_backend_internal_dto.ActivityDTO": {
            "description": "Activity data",
            "type": "object",
            "properties": {
                "date": {
                    "type": "string",
                    "example": "2026-01-04"
                },
                "hours": {
                    "type": "number",
                    "example": 2.5
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ActivityName"
                        }
                    ],
                    "example": "study"
                },
                "note": {
                    "type": "string",
                    "example": "Worked on Go project"
                }
            }
        },
...
        "github_com_aman1117_backend_internal_dto.CreateActivityRequest": {
            "description": "Create or update an activity",
            "type": "object",
            "properties": {
                "activity": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ActivityName"
                        }
                    ],
                    "example": "study"
                },
                "date": {
                    "description": "Format: YYYY-MM-DD",
                    "type": "string",
                    "example": "2026-01-04"
                },
                "hours": {
                    "type": "number",
                    "example": 2.5
                },
                "note": {
                    "type": "string",
                    "example": "Worked on Go project"
                },
                "username": {
                    "type": "string",
                    "example": "john_doe"
                }
            }
        },
        "github_com_aman1117_backend_internal_dto.DataResponse": {
            "description": "Generic response with data payload",
            "type": "object",
            "properties": {
                "data": {},
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "github_com_aman1117_backend_internal_dto.ErrorResponse": {
            "description": "Generic error response",
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                },
                "error_code": {
                    "type": "string",
                    "example": "INVALID_REQUEST"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "github_com_aman1117_backend_internal_dto.GetActivitiesRequest": {
            "description": "Get activities for a date range",
            "type": "object",
            "properties": {
                "end_date": {
                    "description": "Format: YYYY-MM-DD",
                    "type": "string",
                    "example": "2026-01-07"
                },
                "start_date": {
                    "description": "Format: YYYY-MM-DD",
                    "type": "string",
                    "example": "2026-01-01"
                },
                "username": {
                    "type": "string",
                    "example": "john_doe"
                }
            }
        },
...
        "github_com_aman1117_backend_internal_dto.SuccessResponse": {
            "description": "Generic success response",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "models.ActivityName": {
            "type": "string",
            "enum": [
                "sleep",
                "study",
                "book_reading",
                "eating",
                "friends",
                "grooming",
                "workout",
                "reels",
                "family",
                "idle",
                "creative",
                "travelling",
                "errand",
                "rest",
                "entertainment",
                "office"
            ],
            "x-enum-varnames": [
                "ActivitySleep",
                "ActivityStudy",
                "ActivityBookReading",
                "ActivityEating",
                "ActivityFriends",
                "ActivityGrooming",
                "ActivityWorkout",
                "ActivityReels",
                "ActivityFamily",
                "ActivityIdle",
                "ActivityCreative",
                "ActivityTravelling",
                "ActivityErrand",
                "ActivityRest",
                "ActivityEntertainment",
                "ActivityOffice"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}